backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
anthropic==0.28.0
stripe==10.5.0
supabase==2.5.0
PyPDF2==3.0.1
python-docx==1.1.2
reportlab==4.2.2
gunicorn==22.0.0
flask-limiter==3.8.0
requests==2.32.3
PyJWT==2.8.0
